UTOPIC/requirements.txt
tqdm
easydict
h5py
nibabel
ninja
pyyaml
scikit-learn
open3d==0.10.0
numpy==1.22.3
scipy==1.5.0
tensorboard==2.0.0
protobuf==3.19.4

